Refactor/login signup pages 1 (#106)
* feat(components): added AuthLogin component

component to be used in the context of logging in users

* feat(components): added AuthSignup component

component to be used in the context of signing up users

* refactor(pages): added major improvements to login and signup pages
