+++ packages/frontend/src/components/auth/AuthLogin/index.ts
@@ -1,0 +1,3 @@
+import AuthLogin from './AuthLogin';
+
+export default AuthLogin;
