+++ packages/frontend/src/components/auth/AuthSignup/index.ts
@@ -1,0 +1,3 @@
+import NewUserForm from './NewUserForm';
+
+export default NewUserForm;
